J2N.Collections.Generic: Added StringComparerMetadataSerializer to centralize the constants, reading, and writing of known StringComparer types, since it will be useful to keep this uniform for other collection types that support alternate lookup (PR feedback)
